// src/types/userData.ts
export type UserData = {
  id: string;
  nombres: string;
  apellidos: string;
  edad: string;
  cedula: string;
  telefono: string;
  direccion: string;
  sangre: string;
  historia: string;
  nombres_acudiente: string;
  apellidos_acudiente: string;
  telefono_acudiente: string;
  cedula_acudiente: string;
  direccion_acudiente: string;
};